+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+# v2.6.2 Intel® Geti™ SDK (08-01-2024)
+## What's Changed
+* Bugfix: inference not working for classification projects with label containing spaces in their name by @maxxgx in https://github.com/openvinotoolkit/geti-sdk/pull/539
+
+**Full Changelog**: https://github.com/openvinotoolkit/geti-sdk/compare/v2.6.1...v2.6.2
+
+
 # v2.6.1 Intel® Geti™ SDK (02-01-2024)
 ## What's Changed
-* CVS-159908 - Fix empty label ID in configuration [develop] by @maxxgx in https://github.com/openvinotoolkit/geti-sdk/pull/534
+* Bugfix: empty label sometimes not recognized during inference by @maxxgx in https://github.com/openvinotoolkit/geti-sdk/pull/534
 
 ## New Contributors
 * @maxxgx made their first contribution in https://github.com/openvinotoolkit/geti-sdk/pull/534